Remove preload functionality
