+++ src/app.js
@@ -9,7 +9,7 @@ import './core/friendly_errors/fes_core';
 import './core/friendly_errors/sketch_reader';
 import './core/helpers';
 import './core/legacy';
-import './core/preload';
+// import './core/preload';
 import './core/p5.Element';
 import './core/p5.Graphics';
 // import './core/p5.Renderer';
